refactor: post api/hooks 통합
