+++ src/entities/post/api/postApi.ts
@@ -1,8 +1,12 @@
 import {
-  GetPostsRequestType,
   GetPostsResponseType,
+  GetPostsRequestType,
+  PostPostRequestType,
+  PostPostResponseType,
+  PutPostRequestType,
+  PutPostResponseType,
   TagType,
-} from "../model/types";
+} from "@/entities/post/model/types";
 import { createApi } from "@/shared/api/baseApi";
 
 const api = createApi("/api/posts");
@@ -30,5 +34,25 @@ export const postApi = {
   // 태그 목록 조회
   async getTags(): Promise<TagType[]> {
     return api.get<TagType[]>("/tags");
+  },
+
+  // 게시물 추가
+  async createPost(
+    postData: PostPostRequestType,
+  ): Promise<PostPostResponseType> {
+    return api.post<PostPostResponseType>("/add", postData);
+  },
+
+  // 게시물 수정
+  async updatePost(
+    id: number,
+    postData: PutPostRequestType,
+  ): Promise<PutPostResponseType> {
+    return api.put<PutPostResponseType>(`/${id}`, postData);
+  },
+
+  // 게시물 삭제
+  async deletePost(id: number): Promise<void> {
+    return api.delete(`/${id}`);
   },
 };
