--- src/features/comment/api/index.ts
@@ -1,1 +1,0 @@
-export * from "./hooks";
